0033895: fix null surface crash in fixshape
ShapeFix_Shape crashes with a segfault when given a shape that contains
faces without a surface pointer, for example if the shape is tessellated
or triangulated, or contains tessellated / triangulated subshapes.

I added null checks to the `const Handle(Geom_Surface)&` parameter of
the ShapeAnalysis_Surface constructor, and to the analogous parameter
of the Init method, in each case throwing a Standard_NullObject exception.

Now, ShapeFix_Shape::Perform will throw an exception instead of crashing
with a segfault whenever it encounters a face without a surface.

I have CLA 1124.